+++ src/i18n/languages/enGB.ts
@@ -102,6 +102,7 @@ const dictionary: RawTranslationPackage = {
     HEX2OCT: 'HEX2OCT',
     HLOOKUP: 'HLOOKUP',
     HOUR: 'HOUR',
+    HYPERLINK: 'HYPERLINK',
     IF: 'IF',
     IFERROR: 'IFERROR',
     IFNA: 'IFNA',